refactor: extract useHabits hook
Moves all habit tracking state, effects, computed values, and handlers
out of App.jsx into a dedicated useHabits hook. The hook accepts
{ playUISound } and returns all habit-related state and functions
including habitLongPressTimer ref and syncHealthConnectHabitsRef.
